move page title to it's own folder with own utils
